Add `color.py` and `uncolor.py` :art:
`color.py` colors `call`, `push` and `pop` instructions (sets background
color).  It also add the prefix `>>` to`call` instructions and the
number of argument to its parameters (only available if the function
declaration is defined).  This is useful to quickly identify function
calls, their parameters and the calling convention.

`uncolor.py` removes the background color of all the database.  It can
be used to remove the colors added by `color.py`.  It doesn't remove the
prefixes thought.