Add `RC4.py`
It inspect the whole code looking for instructions like
`mov register, offset`
It tries to decode the bytes which start at the `offset` address using
RC4. If the result is a printable string, it adds a comment in that
location with the decoded string and prints a message to the output
window with the decoded string and the address.